finally.

+++ scripts/test-post.sh
@@ -28,8 +28,6 @@ entries=-1
 entries=$(curl --silent "$BASE_URL/?do=atom" | xmllint --encode utf8 --format - | grep --count "<entry>")
 [ $entries -eq 1 ] || { echo "expected exactly one <entry>, found $entries" && exit 4 ; }
 
-rm *.txt
-
 #####################################################
 # Step 1: fetch token to login and add a new link:
 TOKEN=$(curl --get --url "$BASE_URL" \
@@ -39,8 +37,8 @@ TOKEN=$(curl --get --url "$BASE_URL" \
 	--data-urlencode "source=Source Text" \
 	--dump-header head.txt --cookie cook.txt --cookie-jar cook.txt --location --silent \
 | xsltproc --html response.xslt - 2>/dev/null \
-| xmllint --xpath '/shaarli/input[@name="token"]/@value' - \
-| cut -c 8- | tr -d '"')
+| xmllint --xpath 'string(/shaarli/input[@name="token"]/@value)' -)
+# string(..) http://stackoverflow.com/a/18390404
 
 # the precise length doesn't matter, it just has to be significantly larger than ''
 [ $(printf "%s" $TOKEN | wc -c) -eq 40 ] || { echo "expected TOKEN of 40 characters, but found $TOKEN of $(printf "%s" $TOKEN | wc -c)" && exit 5 ; }
@@ -53,6 +51,8 @@ curl --url "${BASE_URL}$(grep -F 'Location: ' head.txt | tr -d '\n' | cut -c 11-
   --data-urlencode "token=$TOKEN" \
   --dump-header head.txt --cookie cook.txt --cookie-jar cook.txt --location --silent \
 | xsltproc --html --output "$tmp".xml response.xslt - 2>/dev/null
+
+[ $(xmllint --xpath 'count(/shaarli/is_logged_in[@value="true"])' "$tmp".xml) -eq 1 ] || { echo "expected to be logged in now" && exit 6 ; }
 
 # turn response.xml form input field data into curl commandline parameters or post file
 ruby response2post.rb < "$tmp".xml > "$tmp".post
@@ -68,12 +68,17 @@ curl --url "${BASE_URL}$(grep -F 'Location: ' head.txt | tr -d '\n' | cut -c 11-
   --dump-header head.txt --cookie cook.txt --cookie-jar cook.txt --location --trace-ascii "$tmp".trace 2>/dev/null \
 | xsltproc --html --output "$tmp".xml response.xslt - 2>/dev/null
 
+cat "$tmp".trace
+
 #####################################################
 # TODO: watch out for error messages like e.g. ip bans or the like.
+
+[ $(xmllint --xpath 'count(/shaarli/is_logged_in[@value="true"])' "$tmp".xml) -eq 1 ] || { echo "expected to be still logged in" && exit 7 ; }
 
 rm "$tmp".post "$tmp".trace "$tmp".xml head.txt cook.txt
 
 # check post-condition - there must be 2 entries now:
 entries=-1
+# 
 entries=$(curl --silent "$BASE_URL/?do=atom" | xmllint --xpath 'count(/*/*[local-name()="entry"])' -)
-[ $entries -eq 2 ] || { echo "expected exactly two <entry>, found $entries" && exit 6 ; }
+[ $entries -eq 2 ] || { echo "expected exactly two <entry>, found $entries" && exit 18 ; }
